Move xmpp models back into core
