+++ src/app/core/components/bread-crumb/bread-crumb.component.ts
@@ -1,7 +1,8 @@
 import {Component, ViewEncapsulation} from '@angular/core';
 import {ActivatedRoute, NavigationEnd, Router} from '@angular/router';
-import {ErrorLogService, XmppService} from '../../index';
-import {BreadCrumb, BreadCrumbs} from './index';
+import {XmppService} from '../../xmpp';
+import {ErrorLogService} from '../../errors';
+import {BreadCrumb, BreadCrumbs} from './bread-crumb';
 import {Observable} from 'rxjs/Observable';
 import 'rxjs/add/operator/distinctUntilChanged';
 import 'rxjs/add/operator/map';
