+++ src/app/core/xmpp/models/affiliation.ts
