+++ src/app/core/xmpp/models/jid-affiliation.ts
