+++ src/app/core/xmpp/models/persisted-item.ts
