+++ src/app/core/xmpp/models/topic.ts
