+++ src/app/core/xmpp/models/xmpp-error-condition.ts
