+++ src/app/models/index.ts
@@ -1,12 +1,13 @@
-export * from './affiliation';
 export * from './config';
-export * from './form-models';
-export * from './jid-affiliation';
 export * from './log-level';
-export * from './persisted-item';
-export * from './subscription';
-export * from './subscription-state';
-export * from './topic';
 export * from './xmpp-error';
-export * from './xmpp-error-condition';
-export * from './xmpp-iq-type';
+
+export * from '../core/xmpp/models/affiliation';
+export * from '../core/xmpp/models/form-models';
+export * from '../core/xmpp/models/jid-affiliation';
+export * from '../core/xmpp/models/persisted-item';
+export * from '../core/xmpp/models/subscription';
+export * from '../core/xmpp/models/subscription-state';
+export * from '../core/xmpp/models/topic';
+export * from '../core/xmpp/models/xmpp-error-condition';
+export * from '../core/xmpp/models/xmpp-iq-type';
